## 펜 그리기 추가 기능 개발 * 변경 파일 다시 적용하여 커밋

+++ wos_intern_project/wos_intern_project/PhotoDrawCanvasView.h
@@ -16,7 +16,7 @@ typedef NS_ENUM(NSInteger, Mode) {
 @interface PhotoDrawCanvasView : UIView
 
 @property (nonatomic, strong) UIColor *lineColor;
-@property (nonatomic, assign) CGFloat lineWidth;
+@property (nonatomic, assign) NSInteger lineWidth;
 @property (nonatomic, assign) NSInteger drawMode;
 
 - (UIImage *)getPathImage;
